add: [email] send message from seleced template by id
