+++ internal/app/apiserver/server.go
@@ -3,32 +3,20 @@ package apiserver
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"github.com/google/uuid"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/sessions"
 	"github.com/nizepart/rest-go/internal/app/email_service"
 	"github.com/nizepart/rest-go/internal/app/store"
-	"github.com/nizepart/rest-go/model"
 	"github.com/sirupsen/logrus"
-	"gopkg.in/gomail.v2"
 	"net/http"
 	"time"
 )
 
 const (
-	sessionName = "mailbomber"
-	ctxKeyUser  = ctxKey(iota)
-	ctxKeyRequestID
+	ctxKeyRequestID = iota
 )
-
-var (
-	errIncorrectEmailOrPassword = errors.New("incorrect email or password")
-	errNotAuthenticated         = errors.New("not authenticated")
-)
-
-type ctxKey int8
 
 type server struct {
 	router        *mux.Router
@@ -74,57 +62,8 @@ func (s *server) configureRouter() {
 
 	template := emails.PathPrefix("/template").Subrouter()
 	template.HandleFunc("/create", s.handleEmailTemplateCreate()).Methods("POST")
-}
-
-func (s *server) handleEmailTemplateCreate() http.HandlerFunc {
-	type request struct {
-		Subject string `json:"subject"`
-		Body    string `json:"body"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		req := &request{}
-		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
-			s.error(w, r, http.StatusBadRequest, err)
-			return
-		}
-
-		et := &model.EmailTemplate{
-			Subject: req.Subject,
-			Body:    req.Body,
-		}
-		if err := s.store.EmailTemplate().Create(et); err != nil {
-			s.error(w, r, http.StatusUnprocessableEntity, err)
-			return
-		}
-
-		s.respond(w, r, http.StatusCreated, et)
-	}
-}
-
-func (s *server) handleEmailSend() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		req := &model.Message{}
-		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
-			s.error(w, r, http.StatusBadRequest, err)
-			return
-		}
-
-		m := gomail.NewMessage()
-		m.SetHeader("From", req.From)
-		m.SetHeader("To", req.To...)
-		m.SetHeader("Cc", req.Cc...)
-		m.SetHeader("Subject", req.Subject)
-		m.SetBody(req.BodyType, req.Body)
-
-		if req.Attach != "" {
-			m.Attach(req.Attach)
-		}
-
-		s.emailService.Send(m)
-
-		s.respond(w, r, http.StatusOK, nil)
-	}
+	template.HandleFunc("/{id}", s.handleEmailTemplateGet()).Methods("GET")
+	template.HandleFunc("/{id}/send", s.handleEmailSend()).Methods("POST")
 }
 
 func (s *server) LogRequest(next http.Handler) http.Handler {
@@ -149,98 +88,6 @@ func (s *server) setRequestID(next http.Handler) http.Handler {
 		w.Header().Set("X-Request-ID", id)
 		next.ServeHTTP(w, r.WithContext(context.WithValue(r.Context(), ctxKeyRequestID, id)))
 	})
-}
-
-func (s *server) authenticateUser(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		session, err := s.sessionsStore.Get(r, sessionName)
-		if err != nil {
-			s.error(w, r, http.StatusInternalServerError, err)
-			return
-		}
-
-		id, ok := session.Values["user_id"]
-		if !ok {
-			s.error(w, r, http.StatusUnauthorized, errNotAuthenticated)
-			return
-		}
-
-		u, err := s.store.User().FindByID(id.(int))
-		if err != nil {
-			s.error(w, r, http.StatusUnauthorized, errNotAuthenticated)
-			return
-		}
-
-		next.ServeHTTP(w, r.WithContext(context.WithValue(r.Context(), ctxKeyUser, u)))
-	})
-}
-
-func (s *server) handleWhoami() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		s.respond(w, r, http.StatusOK, r.Context().Value(ctxKeyUser).(*model.User))
-	}
-}
-
-func (s *server) handleUsersCreate() http.HandlerFunc {
-	type request struct {
-		Email    string `json:"email"`
-		Password string `json:"password"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		req := &request{}
-		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
-			s.error(w, r, http.StatusBadRequest, err)
-			return
-		}
-
-		u := &model.User{
-			Email:    req.Email,
-			Password: req.Password,
-		}
-		if err := s.store.User().Create(u); err != nil {
-			s.error(w, r, http.StatusUnprocessableEntity, err)
-			return
-		}
-
-		u.Sanitize()
-		s.respond(w, r, http.StatusCreated, u)
-	}
-}
-
-func (s *server) handleSessionsCreate() http.HandlerFunc {
-	type request struct {
-		Email    string `json:"email"`
-		Password string `json:"password"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		req := &request{}
-		if err := json.NewDecoder(r.Body).Decode(req); err != nil {
-			s.error(w, r, http.StatusBadRequest, err)
-			return
-		}
-
-		u, err := s.store.User().FindByEmail(req.Email)
-		if err != nil || !u.ComparePassword(req.Password) {
-			s.error(w, r, http.StatusUnauthorized, errIncorrectEmailOrPassword)
-			return
-		}
-
-		session, err := s.sessionsStore.Get(r, sessionName)
-		if err != nil {
-			s.error(w, r, http.StatusInternalServerError, err)
-		}
-
-		session.Values["user_id"] = u.ID
-		if err := s.sessionsStore.Save(r, w, session); err != nil {
-			s.error(w, r, http.StatusInternalServerError, err)
-			return
-		}
-
-		s.respond(w, r, http.StatusOK, nil)
-
-	}
 }
 
 func (s *server) error(w http.ResponseWriter, r *http.Request, code int, err error) {
