+++ model/emailtmplt.go
@@ -3,9 +3,10 @@ package model
 import validation "github.com/go-ozzo/ozzo-validation"
 
 type EmailTemplate struct {
-	ID      int    `json:"id"`
-	Subject string `json:"subject"`
-	Body    string `json:"body"`
+	ID       int    `json:"id"`
+	Subject  string `json:"subject"`
+	Body     string `json:"body"`
+	BodyType string `json:"body_type"`
 }
 
 func (et *EmailTemplate) Validate() error {
@@ -13,5 +14,6 @@ func (et *EmailTemplate) Validate() error {
 		et,
 		validation.Field(&et.Subject, validation.Required, validation.Length(1, 100)),
 		validation.Field(&et.Body, validation.Required),
+		validation.Field(&et.BodyType, validation.Required),
 	)
 }
